Continued development of pokemon battle system
